Support multiple background music songs
Instead of one singletone and hardcoded reference to the Koji background song, you can now add as many sounds as you want, and they get stored in map. The source of the music is now passed to the hook instead of hardcoded inside of it.
Also renamed "App" folder to be "app" (the folder is not a component, but a feature).